src/requirements.txt
abs==0.0.0
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
ale-py==0.8.1
alembic==1.11.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
banal==1.0.6
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.26.4
contourpy==1.1.0
crefi==2.0.9
cycler==0.11.0
dataset==1.6.0
datasets==2.13.0
dateparser==1.2.0
dbms==1.1.1
dill==0.3.6
distlib==0.3.8
doc==0.1.0
docker==6.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Farama-Notifications==0.0.4
fastapi==0.101.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.3.3
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.41.0
gradio_client==0.5.0
greenlet==2.0.2
grpcio==1.58.0
# Editable install with no version control (gym==0.26.2)
-e /home/zhouyuxuan/workspace/pythonWorkspace/bertProject/gym
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
httmock==1.4.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-resources==6.0.1
install==1.3.5
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lit==16.0.6
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mysql-connector-python==8.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.8.0.76
openpyxl==3.1.2
orjson==3.9.5
packaging==23.1
pandas==2.0.2
Pillow==10.0.0
platformdirs==4.2.0
prodict==0.8.18
prometheus-api-client==0.5.4
prometheus-client==0.19.0
protobuf==4.24.3
psutil==5.9.5
psycopg2-binary==2.9.6
pyaml==23.7.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pygame==2.5.1
Pygments==2.16.1
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pyxattr==0.8.1
PyYAML==6.0
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scikit-optimize==0.9.0
scipy==1.11.1
semantic-version==2.10.0
Shimmy==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spaces==0.12.0
spinup==0.1
SQLAlchemy==1.4.48
stable-baselines3==2.1.0
starlette==0.27.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensordict==0.1.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchrl==0.1.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
utils==1.0.1
uvicorn==0.23.2
virtualenv==20.25.1
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.3.7
xxhash==3.2.0
yarl==1.9.2
